Add HLN9820A to the parts list.
This is the Waris (HT750 and HT1250) accessory port dust cover.  The
Waris radios have a screw that holds this cover in place in addition to
the ring that you put around the antenna.  The HLN9820A is the whole
cover, screw and all.

NOTE:  If you try to order it from businessonline.motorola.com, two
things will happen.  First, you'll get hit with a minimum quantity
requirement of 5.  Second, the order won't go through unless you use
Internet Explorer because the credit card processing page URL is
returned to you with backslashes in it.  Ooops.

Best to call the 800 number and order via the phone.  There are no
quantity restrictions that way.