Start section for XPR7550 codeplugs.
The XPR7550 is a handheld MotoTRBO radio, which is a DMR radio.  These
require a different version of CPS and a programming cable that does not
use the RIB.  DMR radios are complicated to program and require lots of
contacts and channel definitions.  My codeplugs include definitions for
the DMR-MARC worldwide network as well as other networks as I find them.
I also keep analog systems programmed in for non-DMR use.